examples working

+++ src/Robot-phy-parameters.cpp
@@ -7,7 +7,7 @@
  * @param inertia A struct that contains all angular inertia values. (Unit: [(Kg * m^2) / degree])
  * @param damping A struct that contains all angular damping values. (Unit: [(N * s) / degree])
  */
-void Robot::setInertiaDamping(std::vector<double> inertia, std::vector<double> damping)
+void Robot::setInertiaDamping(std::vector<float> inertia, std::vector<float> damping)
 {
     AngularInfo commandInertia;
     AngularInfo commandDamping;
@@ -33,7 +33,7 @@ void Robot::setInertiaDamping(std::vector<double> inertia, std::vector<double> d
  *@param inertia A struct that contains all cartesian inertia values. (Translation unit: [Kg] - Orientation unit: [Kg * m^2])
  *@param damping A struct that contains all cartesian damping values. (Translation unit: [(N * s) / m] - Orientation unit: [(N * s) / Rad])
  */
-void Robot::setCartesianInertiaDamping(std::vector<double> inertia, std::vector<double> damping)
+void Robot::setCartesianInertiaDamping(std::vector<float> inertia, std::vector<float> damping)
 {
     CartesianInfo commandInertia;
     CartesianInfo commandDamping;
@@ -52,7 +52,7 @@ void Robot::setCartesianInertiaDamping(std::vector<double> inertia, std::vector<
  * This function sets the delta position over an actuator position limit where the virtual wall begins. The parameters will reset to default values every time the robotic arm is rebooted.
  * @param command The gain of each actuator.
  */
-void Robot::setPositionLimitDistance(std::vector<double>command)
+void Robot::setPositionLimitDistance(std::vector<float>command)
 {
     float* c = new float[command.size()];
     for(unsigned int i = 0; i < command.size(); i++)
@@ -63,7 +63,7 @@ void Robot::setPositionLimitDistance(std::vector<double>command)
 /**
  * This function sets the active spasm filter values of the robotical arm. This function is only used by rehab clients. Regular applications shouldn't use it.
  */
-void Robot::setSpasmFilterValues(std::vector<double>command, int activationStatus)
+void Robot::setSpasmFilterValues(std::vector<float>command, int activationStatus)
 {
     float* c = new float[command.size()];
     for(unsigned int i = 0; i < command.size(); i++)
@@ -75,7 +75,7 @@ void Robot::setSpasmFilterValues(std::vector<double>command, int activationStatu
  * This function set the PID of all actuators.
  * @param newPID PID values
  */
-void Robot::setPID(std::vector<std::vector<double> > newPID) {
+void Robot::setPID(std::vector<std::vector<float> > newPID) {
 	for (int actuatorNumber = 1; actuatorNumber < newPID.size() + 1; actuatorNumber++) {
 		setPID(actuatorNumber, newPID[actuatorNumber - 1]);
 	}
@@ -86,7 +86,7 @@ void Robot::setPID(std::vector<std::vector<double> > newPID) {
  * @param actuatorNumber joint number
  * @param newPID PID values
  */
-void Robot::setPID(int actuatorNumber, std::vector<double> newPID) {
+void Robot::setPID(int actuatorNumber, std::vector<float> newPID) {
 	float P, I, D;
 	unsigned int actuatorPIDAddress;
 
@@ -107,19 +107,19 @@ void Robot::setPID(int actuatorNumber, std::vector<double> newPID) {
  * This function is used to run a sequence to estimate the optimal gravity parameters when the robot is standing (Z). The arm must be in Trajectory-Position mode before this function is called. Before using this procedure, you should make sure that the torque sensors are all well calibrated. This procedure is explained in the user guide and in the Advanced Specification Guide. When the program is launched, the robot will execute a trajectory. The user must remain alert and turn the robot off if something wrong occurs (for example if the robot collides with an object). When the program ends, it will output the parameters in the console and in a text file named “ParametersOptimal_Z.txt” in the program folder. These parameters can then be sent as input to the function SetOptimalZParam().
  * @param type Robot type (MICO_6DOF_SERVICE)
  */
-std::vector<double> Robot::runGravityZEstimationSequence(ROBOT_TYPE type)
+std::vector<float> Robot::runGravityZEstimationSequence(ROBOT_TYPE type)
 {
     (*MySwitchTrajectoryTorque)(POSITION);
     double OptimalzParam[OPTIMAL_Z_PARAM_SIZE];
     (*MyRunGravityZEstimationSequence)(type, OptimalzParam);
-    return std::vector<double>(std::begin(OptimalzParam), std::end(OptimalzParam));
+    return std::vector<float>(std::begin(OptimalzParam), std::end(OptimalzParam));
 }
 
 /**
  * This function sets the gravity parameters using the manual mode. For Mico and Jaco, the input is a vector of 42 floats. The parameters are: [m1,m2,m3,m4,m5,m6,x1,x2,x3,x4,x5,x6,y1,y2,y3,y4,y5,y6,z1,z2,z3,z4,z5,z6] where m is the links mass, and x,y,z is the links center of mass position. The parameters will reset to default values every time the robotic arm is rebooted.
- * @param Gravity parameters vector of size 42
+ * @param command parameters vector of size 42
  */
-void Robot::setGravityManualInputParam(std::vector<double> command)
+void Robot::setGravityManualInputParam(std::vector<float> command)
 {
     if(command.size() != 42)
     {
@@ -139,7 +139,7 @@ void Robot::setGravityManualInputParam(std::vector<double> command)
  * This function is used to set the gravity parameters using the optimal mode. For Mico and Jaco, the input is a vector of 16 floats. For now, this mode only works when the robot is standing (G = [0;0;-9.81]). The parameters will reset to default values every time the robotic arm is rebooted. In order to find these values, the user must run the function RunGravityZEstimationSequence(). The parameters will reset to default values every time the robotic arm is rebooted.
  * @param command the gravity parameters vector
  */
-void Robot::setGravityOptimalZParam(std::vector<double> command)
+void Robot::setGravityOptimalZParam(std::vector<float> command)
 {
     (*MySwitchTrajectoryTorque)(POSITION);
     (*MySetGravityType)(MANUAL_INPUT);
@@ -153,7 +153,7 @@ void Robot::setGravityOptimalZParam(std::vector<double> command)
  * This function sets the payload gravity parameters. The input is a vector of 4 floats : GravityPayload = [pm, px, py, pz]. The parameter pm is the payload mass while px, py and pz are the payload center of mass positions measured from the end-effector frame. The parameters will reset to default values every time the robotic arm is rebooted.
  * @param command the gravity payload vector of size 4
  */
-void Robot::setGravityPayload(std::vector<double> command)
+void Robot::setGravityPayload(std::vector<float> command)
 {
     (*MySwitchTrajectoryTorque)(POSITION);
     (*MySetGravityType)(MANUAL_INPUT);
@@ -173,7 +173,7 @@ void Robot::setGravityPayload(std::vector<double> command)
  * This function sets the gravity vector direction. By default, the gravity is estimated for the classic case where the robot is standing: G = [0;0;-9.81]. The parameters will reset to default values every time the robotic arm is rebooted.
  * @param command The gravity vector of size 3.
  */
-void Robot::setGravityVector(std::vector<double> command)
+void Robot::setGravityVector(std::vector<float> command)
 {
     (*MySwitchTrajectoryTorque)(POSITION);
     (*MySetGravityType)(MANUAL_INPUT);
